We now add 0 to 0

+++ Kata/Fractions/spec/sample.spec.ts
@@ -1,7 +1,7 @@
-import {Sample} from "../src/sample"
+import {Fraction} from "../src/sample";
 describe("AddFractions", ()=> {
     it("should return zero + zero", ()=> {
         var sum: Fraction = new Fraction(0).plus(new Fraction(0));
-        expect(sum).toBe(0);
+        expect(sum.intValue()).toBe(0);
     });
 });
